requirements.txt
astroid==1.6.1
autopep8==1.3.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.9
Flask==0.12.2
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pycodestyle==2.3.1
pylint==1.8.2
python-dotenv==0.7.1
requests==2.18.4
six==1.11.0
spotipy==2.4.4
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
